--- packages/client/src/store/discardedCard.types.ts
@@ -1,5 +1,0 @@
-import { CardType } from '../pages/game/types'
-
-export interface DiscardedCardSlice {
-  discardedCard: CardType | null
-}
